ekump/APMSP-1279 benchmark trace exporting (#531)
* Extract logic to build TracerPayloadCollection from sidecar into TracerPayload.

This functionality to turn msgpack bytes and header tags into a
TracerPayloadCollection should live somewhere more common than the sidecar as it will need to
also be used by data-pipeline (and presumably agentless). It only supports v04 at the moment, but
should eventually support v05 and v07 in follow up PRs.

* Adding some initial benchmarks for deserializing traces from msgpack

Deserializing msgpack bytes into an internal representation is currently
inefficient. Before work starts on fixing the performance issues we want
to set some baseline benchmarks.

* Switch the chunk processing to use a trait impl instead of dyn
The TraceChunk processor that is used primarily by the mini-agent (but is also created and passed by Sidecar and data-pipeline) was previously a closure that was dynamically dispatched. This was changed to be a trait in order to remove the dynamic-ness and should improve performance and in many cases the compiler should optimize the calls to it away as it is a noop for the Sidecar and data pipeline.


Co-authored-by: Bob Weinand <bob.weinand@datadoghq.com>

---------

Co-authored-by: Bob Weinand <bob.weinand@datadoghq.com>