fix(alloc): potentially dangling temporary (#772)
* fix: potentially dangling temporary

I was trying out the Rust 2024 edition which required a new nightly.
There's now a warning for dangling_pointers_from_temporaries. This
could be a false positive but I am not certain.

* test: skip problematic miri tests on MacOS

These are all due to kqueue, example:

```
error: unsupported operation: can't call foreign function `kqueue` on OS `macos`
   --> /levi.morrison/.cargo/registry/src/index.crates.io-6f17d22bba15001f/mio-1.0.2/src/sys/unix/selector/kqueue.rs:76:48
    |
76  |         let kq = unsafe { OwnedFd::from_raw_fd(syscall!(kqueue())?) };
    |                                                ^^^^^^^^^^^^^^^^^^ can't call foreign function `kqueue` on OS `macos`
```

* test: fix env test when running locally

If the local env has this set, it fails.